+++ examples/main/src/vite-env.d.ts
@@ -1,7 +1,7 @@
 /// <reference types="vite/client" />
 
 declare module 'virtual:wllama-compat' {
-  import type { WllamaCompat } from '@wllama/wllama';
+  import type { WllamaCompat } from 'wllama64';
   const config: WllamaCompat | 'default';
   export default config;
 }